More documentation and some refactoring
Added function and module headers for tinyweb.[hc], request.[hc] and
response.[hc].

Restructured request.c a little.  It now uses the built-in functions
strstr() and strchr() instead of our homebrew find_next_clrf() and
find_next_char() functions.

Changed the comment style in log.[hc] and http.[hc] to use /**/-style
comments instead of //-style comments, as the latter are not ANSI-C
compatible and the change doesn't cost much.

Added Makefile rules to generate Doxygen documentation

+++ src/response.h
@@ -50,7 +50,7 @@ typedef struct {
                                            is generated */
     time_t last_modified;             /*!< The time the requested file was last
                                            modified */
-    unsigned int content_length;      /*!< The length of the requested file in
+    size_t content_length;            /*!< The length of the requested file in
                                            bytes */
     http_content_type_t content_type; /*!< The content type of the requested
                                            type */
